+++ src/init.c
@@ -13,6 +13,7 @@ extern SEXP _later_execCallbacks();
 extern SEXP _later_idle();
 extern SEXP _later_execLater(SEXP, SEXP);
 extern SEXP _later_saveNframesCallback(SEXP);
+extern SEXP _later_next_op_secs();
 
 static const R_CallMethodDef CallEntries[] = {
   {"_later_ensureInitialized",   (DL_FUNC) &_later_ensureInitialized,   0},
@@ -20,6 +21,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"_later_idle",                (DL_FUNC) &_later_idle,                0},
   {"_later_execLater",           (DL_FUNC) &_later_execLater,           2},
   {"_later_saveNframesCallback", (DL_FUNC) &_later_saveNframesCallback, 1},
+  {"_later_next_op_secs",        (DL_FUNC) &_later_next_op_secs,        0},
   {NULL, NULL, 0}
 };